+++ src/components/Icons/index.ts
@@ -1,31 +1,31 @@
 // Social media brands
-export { default as IconDribbble } from './social/Dribbble.Icon';
-export { default as IconFacebook } from './social/Facebook.Icon';
-export { default as IconInstagram } from './social/Instagram.Icon';
-export { default as IconTwitter } from './social/Twitter.Icon';
-export { default as IconLinkedIn } from './social/LinkedIn.Icon';
-export { default as IconMedium } from './social/Medium.Icon';
-export { default as IconGithub } from './social/Github.Icon';
-export { default as IconUnsplash } from './social/Unsplash.Icon';
-export { default as IconYouTube } from './social/YouTube.Icon';
-export { default as IconPatreon } from './social/Patreon.Icon';
-export { default as IconPaypal } from './social/Paypal.Icon';
-export { default as IconWeibo } from './social/Weibo.Icon';
+export { default as IconDribbble } from './social/IconDribbble';
+export { default as IconFacebook } from './social/IconFacebook';
+export { default as IconInstagram } from './social/IconInstagram';
+export { default as IconTwitter } from './social/IconTwitter';
+export { default as IconLinkedIn } from './social/IconLinkedIn';
+export { default as IconMedium } from './social/IconMedium';
+export { default as IconGithub } from './social/IconGithub';
+export { default as IconUnsplash } from './social/IconUnsplash';
+export { default as IconYouTube } from './social/IconYouTube';
+export { default as IconPatreon } from './social/IconPatreon';
+export { default as IconPaypal } from './social/IconPaypal';
+export { default as IconWeibo } from './social/IconWeibo';
 
 // UI
-export { default as IconChevronLeft } from './ui/ChevronLeft.Icon';
-export { default as IconCopied } from './ui/Copied.Icon';
-export { default as IconCopy } from './ui/Copy.Icon';
-export { default as IconEx } from './ui/Ex.Icon';
-export { default as IconLink } from './ui/Link.Icon';
-export { default as IconToggleOpen } from './ui/ToggleOpen.Icon';
-export { default as IconToggleClose } from './ui/ToggleClose.Icon';
-export { default as IconRows } from './ui/Rows.Icon';
-export { default as IconTiles } from './ui/Tiles.Icon';
-export { default as IconDone } from './ui/Done.Icon';
-export { default as IconExternalLink } from './ui/ExternalLink.Icon';
-export { default as IconHashLink } from './ui/HashLink.Icon';
-export { default as IconHeart } from './ui/Heart.Icon';
-export { default as IconHandclap } from './ui/Handclap.Icon';
-export { default as Arrow } from './ui/Arrow.Icon';
-export { default as IconNotFound } from './ui/404.Icon';
+export { default as IconChevronLeft } from './ui/IconChevronLeft';
+export { default as IconCopied } from './ui/IconCopied';
+export { default as IconCopy } from './ui/IconCopy';
+export { default as IconEx } from './ui/IconEx';
+export { default as IconLink } from './ui/IconLink';
+export { default as IconToggleOpen } from './ui/IconToggleOpen';
+export { default as IconToggleClose } from './ui/IconToggleClose';
+export { default as IconRows } from './ui/IconRows';
+export { default as IconTiles } from './ui/IconTiles';
+export { default as IconDone } from './ui/IconDone';
+export { default as IconExternalLink } from './ui/IconExternalLink';
+export { default as IconHashLink } from './ui/IconHashLink';
+export { default as IconHeart } from './ui/IconHeart';
+export { default as IconHandclap } from './ui/IconHandclap';
+export { default as Arrow } from './ui/IconArrow';
+export { default as IconNotFound } from './ui/Icon404';
