+++ src/components/Icons/social/IconLinkedIn.tsx
