+++ src/components/Icons/ui/IconCopied.tsx
