+++ src/components/Icons/ui/IconExternalLink.tsx
