+++ src/components/Icons/ui/IconHandclap.tsx
