+++ src/components/Icons/ui/IconShare.tsx
