+++ src/components/Icons/ui/IconToggleClose.tsx
