+++ src/components/Icons/ui/IconToggleOpen.tsx
