+++ src/components/Image/index.ts
@@ -1,3 +1,3 @@
 export { default } from './Image';
-export { default as ImageZoom } from './Image.Zoom';
-export { default as ImagePlaceholder } from './Image.Placeholder';
+export { default as ImageZoom } from './ImageZoom';
+export { default as ImagePlaceholder } from './ImagePlaceholder';
