+++ src/components/Navigation/index.ts
@@ -1,1 +1,1 @@
-export { default } from '@components/Navigation/Navigation.Header';
+export { default } from '@components/Navigation/NavigationHeader';
