--- src/components/Transitions/index.ts
@@ -1,7 +1,0 @@
-import CSSFadeIn from '@components/Transitions/Transitions.CSS.FadeIn';
-
-export default {
-  CSS: {
-    FadeIn: CSSFadeIn,
-  },
-};
