+++ src/templates/ArticlesTemplate.tsx
@@ -5,7 +5,7 @@ import styled from '@emotion/styled';
 
 import Section from '@components/Section';
 import SEO from '@components/SEO';
-import Paginator from '@components/Navigation/Navigation.Paginator';
+import Paginator from '@components/Navigation/NavigationPaginator';
 
 import ArticlesHero from '@sections/articles/ArticlesHero';
 import ArticlesList from '@sections/articles/ArticlesList';
