+++ src/templates/AuthorTemplate.tsx
@@ -5,7 +5,7 @@ import styled from '@emotion/styled';
 
 import Section from '@components/Section';
 import SEO from '@components/SEO';
-import Paginator from '@components/Navigation/Navigation.Paginator';
+import Paginator from '@components/Navigation/NavigationPaginator';
 
 import AuthorHero from '@sections/author/AuthorHero';
 import AuthorArticles from '@sections/author/AuthorArticles';
